cardConfigData collection moved to Card component
